Fully qualify function calls
Mainly a micro optimisation as a non qualified function name prevents
PHP from knowing which function it needs to call until runtime where it
will do a lookup in the namespace before falling back to the global functions.

Using fully qualified names allows the PHP VM to generate optimized opcodes
for the load of arguments and use an opcode for the array_key_exists() function.